修改APSI代码

+++ mpc4j-s2pc-pir/src/main/java/edu/alibaba/mpc4j/s2pc/pir/PirUtils.java
@@ -444,7 +444,7 @@ public class PirUtils {
      * @param input input.
      * @return bit length.
      */
-    public static int getBitLength(int input) {
+    public static int getBitLength(long input) {
         int count = 0;
         while (input != 0) {
             count++;
